Display `SocialButtons` (closes #2)
I couldn't find an svg for the modern osu! logo so I made one myself using the assets on the osu! website
I've switched the background colour and the colour of the "A" in the AniList logo so it would look better here
I've made this Last.fm logo myself, using its old square logo and the colour of the current favicon